Split cli/commands/misc.ts into six per-verb modules
